\name{SRBreak-package}
\alias{SRBreak-package}
\alias{SRBreak}
\docType{package}
\title{
Identify breakpoints for copy-number variable regions
}
\description{
More about what it does (maybe more than one line)

}
\details{
\tabular{ll}{
Package: \tab SRBreak\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-11-20\cr
License: \tab What license is it under?\cr
}

}
\author{
Who wrote it

Maintainer: Who to complain to <yourfault@somewhere.net>

}
